+++ Source/PropertyTools.Wpf/DataGrid/DataGrid.cs
@@ -240,7 +240,7 @@ namespace PropertyTools.Wpf
             nameof(ControlFactory),
             typeof(IDataGridControlFactory),
             typeof(DataGrid),
-            new UIPropertyMetadata(new DataGridControlFactory()));
+            new UIPropertyMetadata(new DataGridControlFactory(), (d, e) => ((DataGrid)d).UpdateGridContent()));
 
         /// <summary>
         /// Identifies the <see cref="CellDefinitionFactory"/> dependency property.
@@ -249,7 +249,7 @@ namespace PropertyTools.Wpf
             nameof(CellDefinitionFactory),
             typeof(ICellDefinitionFactory),
             typeof(DataGrid),
-            new UIPropertyMetadata(new CellDefinitionFactory()));
+            new UIPropertyMetadata(new CellDefinitionFactory(), (d, e) => ((DataGrid)d).UpdateGridContent()));
 
         /// <summary>
         /// Identifies the <see cref="CurrentCell"/> dependency property.
@@ -2740,13 +2740,22 @@ namespace PropertyTools.Wpf
         /// <summary>
         /// Removes the current editor control.
         /// </summary>
-        private void RemoveEditControl()
+        /// <param name="updateTextBindingSource">
+        /// if set to <c>true</c>, updates the source binding for text editors before removal.
+        /// </param>
+        private void RemoveEditControl(bool updateTextBindingSource = true)
         {
             if (this.currentEditControl != null/* && this.currentEditControl.Visibility == Visibility.Visible*/)
             {
                 var textEditor = this.currentEditControl as TextBox;
                 if (textEditor != null)
                 {
+                    if (updateTextBindingSource)
+                    {
+                        var textBinding = textEditor.GetBindingExpression(TextBox.TextProperty);
+                        textBinding?.UpdateSource();
+                    }
+
                     textEditor.PreviewKeyDown -= this.TextEditorPreviewKeyDown;
                 }
 
@@ -4028,7 +4037,7 @@ namespace PropertyTools.Wpf
                     break;
                 case Key.Escape:
                     BindingOperations.ClearBinding(this.currentEditControl, TextBox.TextProperty);
-                    this.RemoveEditControl();
+                    this.RemoveEditControl(false);
                     e.Handled = true;
                     break;
             }